Making NanoMeaner produce DataIDName
Changed codes so that NanoMeaner the DataIDName produced in NanoMeaner will also be the same as that used in NanoDataSave, saving people time. The other codes have been adapted around for this but have not been tested. If running in Create or not meaned (= nan) then it will ask for the user to write DataIDName