Reworked Agilent data
Made a new creator thing called Import which then works of functions which just save the end data. Also updated the equivalent of the loader i.e. NanoDataLoader to be more flexible, I want to rename it to NanoPlotter. Next step is to make it possible to pre-assess the data to load so one can change the bin width or maximum depth.