Starting to fix issues with previous versions
Updating the code by making a general import function that should be able to work on either a spreadsheet or txt. And it should be able to work for any method and machine.
Essentially it determines whether it is Agilent or Bruker, then depending on the answer it then works them out differently. Crucially however it will not mean the data, as I think I will create a seperate function for this, that way the code is cleaner.
Could have an option that to import all indents found in a spreadsheet, giving them automatically generatred names; OR the user can manually see and reject what indents to do.